On-going test, after RabbitMQ re-installed.
Default server node name is 'rabbit@ASUS' and the '@' part may be why a remote connection is problematic; tested on 'localhost' instead.
Reason for earlier 'hanging' behaviour still unknown but could well be a Windows issue.
"multiprocessing" abandoned as timeout handler in favour of "stopit", as former could not pickle the relevant callback!