Work In Progress: attempt to use ASYNCHRONOUS version of "Publisher Confirms" but it doesn't seem to be viable.
The RabbitMQ advice given for this extension is to update a count of each message as it is sent and add that value to a 'tracking set'.
However in the event of a crash, the kombu/py-amqp "delivery tag counter" would be reset and so there could be a duplication
(or collapse) of counter/delivery_tag values in the tracking set, which would result in confusion. IMO, the 'publish' method should return the
'delivery info' details of the message just sent but even that may not be sufficient to resolve the problem; the delivery_tag count would also need
to be persistent and recoverable.

In summary, Python AMQP libraries (not just Kombu) do not seem to properly support RabbitMQ's "Publisher Confirms/Acknowledgements"
yet, for asyncronous operation. The hope is that the 'confirm_publish' setting of a Connection "transport option" will result in a blocking
version of 'basic_publish', as advertised.